requirements.txt
albumentations==1.3.1
easydict==1.9
matplotlib==3.7.2
numpy==1.26.4
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==2.0.2
Pillow==9.3.0
Pillow==9.5.0
Pillow==10.2.0
scikit_learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
tensorlayerx==0.5.8
torch==2.0.1
torch==1.13.0
torchsummary==1.5.1
torchvision==0.15.2
torchvision==0.14.0
tqdm==4.65.0
